add / custom block style
