+++ src/admin/custom-block-style/item/property-label.js
