+++ src/admin/custom-block-style/item/title-area/delete-button/index.js
